Re-arrange the modules to keep backwards compatibility with master.
Mutable buffers are now the default again, and immutable buffers are
moved to Node.Buffer.Immutable
