+++ test/Test/Main.purs
@@ -3,9 +3,9 @@ module Test.Main where
 import Prelude
 import Effect (Effect)
 import Test.Node.Buffer as Buffer
-import Test.Node.Buffer.Mutable as Mutable
+import Test.Node.Buffer.Immutable as Immutable
 
 main :: Effect Unit
 main = do
   Buffer.test
-  Mutable.test
+  Immutable.test
